transaction post

+++ data/src/main/kotlin/com/diegoferreiracaetano/data/remote/transaction/TransactionEntity.kt
@@ -4,8 +4,8 @@ import com.diegoferreiracaetano.data.remote.user.UserEntity
 import com.google.gson.annotations.SerializedName
 
 internal data class TransactionEntity (
-    @SerializedName("id") val id: Int,
-    @SerializedName("timestamp") val timestamp: Int,
+    @SerializedName("id") val id: Long,
+    @SerializedName("timestamp") val timestamp: Long,
     @SerializedName("value") val value: Float,
     @SerializedName("success") val success: Boolean,
     @SerializedName("status") val status: String,
